Moved core request & response generics to Models namespace
