+++ ZendeskSell/Models/ZendeskSellRequest.cs
@@ -3,11 +3,20 @@ using System.Collections.Generic;
 using System.Text;
 using Newtonsoft.Json;
 
-namespace ZendeskSell
+namespace ZendeskSell.Models
 {
     public class ZendeskSellRequest<T>
         where T : class
     {
+        public ZendeskSellRequest()
+        {
+
+        }
+
+        public ZendeskSellRequest(T data) {
+            Data = data;
+        }
+
         [JsonProperty("data")]
         public T Data { get; set; }
     }
